New continuation line error E127 unnecessary backslash
The continuation backslash is only required if there is not an open
bracket enclosing the end of line.  Make it an error to have it there
anyway.  This broadens the continuation line indenting test to test more
than just indenting, but it is a convenient place to put the test.