Allow tokens following multi–line string literals to avoid E12*
Multi-line literal strings are a little bit of an outsider when it comes
to continuation lines.  If such a token is used, consider the code to be
OK so long as the next token starts either immediately, or at the same
place you'd expect tokens to start again if the multi–line literal
wasn't there.  This is a corner case, none of the examples are
particularly readable, but it shows what does work.