Make E127/E128 stricter about what defines a new level
The current code which permits tokens to be indented to match tokens on
previous lines was a bit too permissive; any token which happened to
line up with any previous token at all on the previous line was allowed;
when really, this only makes sense for homogenous tokens.  Make the
first step towards tidying up this rule; this first part makes some
forms spotted by myint illegal, while still allowing the continuation of
multi-line strings with an additional indent.

There is one example which sees some fallout, which is really a
generalization of the rule which lets multi-line strings continue on the
next line: tokens should be allowed to line up with an identical token
on the previous line.  This generalization will follow.