Relax E127/E128 for aligned homogenous tokens
When using visual indenting, there is a tendency to communicate through
interpretive dance.  Extra whitespace is added to visually communicate
something which might be lost if more rigid rules were enforced.  The
examples in the test suite all start with a token which matches the same
token on the previous line.  So, this new rule permits indents to a
matching level as a token on the previous line, but only if the tokens are
the same.  This allows an example to pass which uses chained method calls,
with each continuation line aligned by the ".".