Relax E127/E128 for aligned homogenous tokens
When using visual indenting, there is a tendency to communicate through
'interpretive dance', which is an affectionate name I have for arbitrary
extra whitespace inserted to visually communicate 'something' to the
reader.

The examples in the test suite all start with a token which matches the
same token on the previous line.  This new rule permits indents to a
matching level as a token on the previous line, but only if the tokens
are the same.

Add some tests to show the quirks with the current rule, which allows
the first visual indent line to align with any token whatsoever.