Fix comments related to unconsistent commits
Comment was made that checks should be ignored by
default. This is now the case and code/documentation
and tests have been updated.

Also, Travis commented that a few things went wrong
on pretty much all Python versions. I think I have
fixed most of the issues (related to the fact that :
 - some AST elements have changed
 - sys.version_info is not a namedtuple on Python 2.6).

Also, I realised that a test would fail as a SyntaxError
only on Python 2.6. For that reason, I've made the
version tag on error somewhat more generic. Now, it can
be any prefix for a version : 2 corresponds to all Python
2 version as 2.6 correponds to 2.6 and so on...