Attach the original source to the compiled tree
Issue:

When flake8 is being piped a file, stdin is consumed by pep8. A checker is
unable to evaluate the original source.

`seek(0)` is not possible since it is illegal in a pipe.  Constructing the
source from the tree gives unreliable line numbers.

The goal here is to build an IDE plugin, working specially with Atom.