Do not report E302 occurring on first logical line
67b8be8 introduced an exception for E302 if the
first top level function or class was preceded
with only one comment line.

PEP 263 explicitly allows for two lines of comments
at the top of a module, in order to support a
hash-bang and an encoding declaration.

When the first top level function or class appears on the
first logical line, i.e. preceded only by comments, E302
should not be reported as the pep8 utility typically
ignores comments appearing anywhere in the code being checked.

Fixes #412