Identify binary operators used as unary operators
Previously we handled the case where binary operators were being used
as unary operators except in the case where they followed another
binary operator, e.g.,

    foo = (1 +
           -10 *
           2 /
           -5)

This change updates the check for W503 to track the previous
non-whitespace token type and token text to check if it is in fact also
a binary operator (because you should never have two binary operators in
a row).

This does not handle invalid syntax, e.g.,

    foo = (bar /
           /baz)

But the false-positive generated for other cases was more harmful than
not catching what will instead be caught by the interpreter.

Closes gh-484