Cut run_check
`run_check` works by iterating over a list of strings, then
dynamically accessing the appropriate checker attributes. Dynamic
attribute access is slow, and doing it in a tight loop is really slow.
Speed can be improved significantly by passing the checker in as an
argument to the check and calling the attributes directly. This
requires some argument shuffling.