Correctly allow *two* blank lines after a block of one-liners.
Previously
```
def oneliner(): pass
def otheroneliner(): pass

def really_long_func():
    with_some_contents
```
would raise an "E302: expected 2 blank lines, found zero" at the last
line of the one liner.  Ultimately, this is due to `expand_indent` being
passed a line whose contents are just a newline and nothing else, and
`expand_indent` thinking that the line is indented by 1 character (the
newline), which is wrong.

Fix that by just stripping the newline, and modify a test to cover this
case.