introduce root argument
The following commit introduces a new argument `--root`, which allows a
caller to specify the base location where a project-specific
configuration can be found. This is primarily to enable callers which
feed data through stdin to specify a possible project configuration
without explicitly configuring the working directory when pycodestyle is
invoked.

Signed-off-by: James Knight <james.d.knight@live.com>